Fixed Hm-2.15 (devblog 1)
